# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentGuard Policy Engine API
  description: |
    **AgentGuard** is a runtime security platform for AI agents. It provides policy-based evaluation,
    audit trails, PII detection, MCP server governance, agent hierarchy management, and OWASP compliance reporting.

    ## Authentication

    AgentGuard uses three authentication schemes:

    - **API Key** (`x-api-key` header): Your tenant API key, prefixed `ag_live_`. Used for most authenticated endpoints.
    - **Agent Key** (`x-api-key` header): Agent-scoped key, prefixed `ag_agent_`. Used for evaluate calls from specific agents.
    - **Admin Key** (`x-admin-key` header or `x-api-key`): Platform admin key. Required for admin-only endpoints.

    ## Rate Limits

    - Unauthenticated: 10 requests/min per IP
    - Authenticated: 100 requests/min per IP
    - Signup: 5 per hour per IP

    ## Base URL

    Production: `https://api.agentguard.tech/api/v1`
  version: 0.10.0
  contact:
    name: AgentGuard Support
    url: https://agentguard.tech
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agentguard.tech/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Tenant API key (prefixed `ag_live_`) or agent key (prefixed `ag_agent_`)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Platform admin key (set via ADMIN_KEY environment variable)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token (future use)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Invalid API key"
        details:
          type: array
          items:
            type: string
          description: Additional error details (validation errors)
...
paths:
...
  /signup:
    post:
      tags: [Auth]
      summary: Create tenant account
      description: |
        Create a new tenant account and receive an API key. The API key is shown **only once** — store it securely.
        Rate limited to 5 signups per hour per IP.
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                  maxLength: 200
                  description: Organization or user name
                  example: "Acme Corp"
                email:
                  type: string
                  format: email
                  description: Contact email address
                  example: "security@acme.com"
            example:
              name: "Acme Corp"
              email: "security@acme.com"
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                    format: uuid
                  apiKey:
                    type: string
                    description: Your API key — store this securely, not shown again
                    example: "ag_live_a1b2c3d4e5f6..."
                  dashboard:
                    type: string
                    format: uri
                  message:
                    type: string
              example:
                tenantId: "550e8400-e29b-41d4-a716-446655440000"
                apiKey: "ag_live_a1b2c3d4e5f67890abcd1234"
                dashboard: "https://app.agentguard.tech"
                message: "Account created. Store your API key securely — it will not be shown again."
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid email format"
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /keys/rotate:
    post:
      tags: [Auth]
      summary: Rotate API key
      description: Generate a new API key and invalidate the current one. The new key is shown only once.
      operationId: rotateKey
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: New API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey:
                    type: string
                    description: New API key — store securely
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── EVALUATE ──────────────────────────────────────────────────────────────
  /evaluate:
    get:
      tags: [Evaluate]
      summary: Evaluate endpoint info
      description: Returns a description of the evaluate endpoint and usage instructions. No authentication required.
      operationId: getEvaluateInfo
      security: []
      responses:
        '200':
          description: Endpoint info
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoint:
                    type: string
                  method:
                    type: string
                  description:
                    type: string
                  requiredFields:
                    type: array
                    items:
                      type: string
                  example:
                    type: object
    post:
      tags: [Evaluate]
      summary: Evaluate an agent action
      description: |
        The core evaluation endpoint. Submits an agent action (tool call) to the policy engine
        and receives a decision: `allow`, `block`, `monitor`, or `require_approval`.

        Authentication is optional — unauthenticated calls use the default policy.
        Authenticated calls use tenant-specific custom policies.
      operationId: evaluate
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool]
              properties:
                tool:
                  type: string
                  maxLength: 200
                  pattern: '^[a-zA-Z0-9_.\-:]+$'
                  description: Tool name being requested
                  example: "file_read"
                params:
                  type: object
                  description: Tool parameters (key-value pairs)
                  example:
                    path: "/etc/config.yaml"
                agentId:
                  type: string
                  description: Agent identifier for scoped evaluation
                sessionId:
                  type: string
                  description: Session identifier for grouping audit events
                context:
                  type: object
                  properties:
                    userId:
                      type: string
                    environment:
                      type: string
                      example: "production"
                messageHistory:
                  type: array
                  description: Recent conversation history for context-aware detection
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, system]
                      content:
                        type: string
            examples:
              allow_example:
                summary: File read (typically allowed)
                value:
                  tool: "file_read"
                  params:
                    path: "/var/log/app.log"
                  agentId: "my-agent-001"
              block_example:
                summary: Shell exec (typically blocked)
                value:
                  tool: "shell_exec"
                  params:
                    command: "rm -rf /tmp"
                  agentId: "my-agent-001"
              approval_example:
                summary: Sensitive DB query (requires approval)
                value:
                  tool: "db_query"
                  params:
                    table: "users"
                  agentId: "my-agent-001"
      responses:
        '200':
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDecision'
              examples:
                allow:
                  value:
                    result: allow
                    riskScore: 10
                    reason: "Matched rule: allow-read-operations"
                    matchedRuleId: "rule-allow-read"
                    durationMs: 3.2
                    approvalId: null
                block:
                  value:
                    result: block
                    riskScore: 500
                    reason: "Shell execution is blocked by security policy"
                    matchedRuleId: "rule-block-shell"
                    durationMs: 2.8
                    approvalId: null
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /usage:
    get:
      tags: [Analytics]
      summary: Usage statistics
      description: |
        Retrieve evaluation usage statistics for the authenticated tenant.
        Includes total evaluations, breakdowns by result, last 24h count,
        top blocked tools, and average response time.
      operationId: getUsageStats
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  totalEvaluations:
                    type: integer
                  blocked:
                    type: integer
                  allowed:
                    type: integer
                  monitored:
                    type: integer
                  requireApproval:
                    type: integer
                  last24h:
                    type: integer
                    description: Evaluations in the last 24 hours
                  topBlockedTools:
                    type: array
                    items:
                      type: object
                      properties:
                        tool:
                          type: string
                        count:
                          type: integer
                  avgResponseMs:
                    type: number
                    description: Average evaluation response time in milliseconds
              example:
                tenantId: "550e8400-e29b-41d4-a716-446655440000"
                totalEvaluations: 1250
                blocked: 42
                allowed: 1100
                monitored: 95
                requireApproval: 13
                last24h: 87
                topBlockedTools:
                  - tool: "shell_exec"
                    count: 25
                  - tool: "eval_code"
                    count: 12
                avgResponseMs: 3.45
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── FEEDBACK ─────────────────────────────────────────────────────────────
...
  /templates/{id}:
    get:
      tags: [Templates]
      summary: Get policy template
      description: Retrieve a specific policy template by its ID.
      operationId: getTemplate
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Template ID (e.g., "owasp-agentic", "soc2-starter", "eu-ai-act", "financial-services", "apra-cps234")
          example: "owasp-agentic"
      responses:
        '200':
          description: Template details with full rule set
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    allOf:
                      - $ref: '#/components/schemas/TemplateObject'
                      - type: object
                        properties:
                          rules:
                            type: array
                            items:
                              $ref: '#/components/schemas/PolicyRule'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /templates/{id}/apply:
    post:
      tags: [Templates]
      summary: Apply policy template
      description: |
        Apply a policy template to this tenant's configuration. This records the application
        in the audit trail and provides the rules for integration into your policy engine.
      operationId: applyTemplate
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "owasp-agentic"
      responses:
        '200':
          description: Template applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  applied:
                    type: boolean
                  templateId:
                    type: string
                  templateName:
                    type: string
                  rulesInTemplate:
                    type: integer
                  message:
                    type: string
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── KILL SWITCH ──────────────────────────────────────────────────────────
  /killswitch:
    get:
      tags: [Kill Switch]
      summary: Get kill switch status
      description: Returns the current state of the global platform kill switch.
      operationId: getKillswitchStatus
      security: []
      responses:
        '200':
          description: Kill switch state
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                  activatedAt:
                    type: string
                    format: date-time
                    nullable: true
    post:
      tags: [Kill Switch]
      summary: Toggle tenant kill switch
      description: |
        Activate or deactivate the tenant-level kill switch.
        When active, all evaluations for this tenant return `block` immediately.
        This is an emergency stop mechanism for incident response.
      operationId: toggleKillswitch
      security:
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: boolean
                  description: Set to true to activate, false to deactivate. Omit to toggle.
            example:
              active: true
      responses:
        '200':
          description: Kill switch state updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: string
                  active:
                    type: boolean
                  activatedAt:
                    type: string
                    format: date-time
                    nullable: true
                  message:
                    type: string
              example:
                tenantId: "550e8400-e29b-41d4-a716-446655440000"
                active: true
                activatedAt: "2024-01-15T10:30:00Z"
                message: "Tenant kill switch ACTIVATED — your evaluations will return BLOCK"

  # ── MCP ADMIT ────────────────────────────────────────────────────────────
...
  /signup/recover:
    post:
      tags: [Auth]
      summary: Recover account by email
      description: |
        Issue a new API key by verifying email ownership. Invalidates all previous keys.
        Rate limited to 2 recovery attempts per hour per IP.
        Returns a generic 404 when the email is not found to avoid enumeration.
      operationId: recoverAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "security@acme.com"
      responses:
        '200':
          description: New API key issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey:
                    type: string
                    description: New API key — previous key invalidated
                  message:
                    type: string
        '400':
          description: Email required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...